Notifications detached from core module handler
